+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New option `--cache-root` and support for the `RUBOCOP_CACHE_ROOT` environment variable. Both can be used to override the `AllCops: CacheRootDirectory` config, especially in a CI setting. ([@sascha-wolf][])
 * [#8582](https://github.com/rubocop-hq/rubocop/issues/8582): Add new `Layout/BeginEndAlignment` cop. ([@koic][])
+* Add new `Lint/PreferAllowlistDenylist` cop. ([@bjohns2][])
 
 ### Bug fixes
 
@@ -4855,3 +4856,4 @@
 [@jaimerave]: https://github.com/jaimerave
 [@Skipants]: https://github.com/Skipants
 [@sascha-wolf]: https://github.com/sascha-wolf
+[@bjohns2]: https://github.com/bjohns2